WIP: ingest blocks and save them in database accordingly

+++ chain-exporter/schema/block.go
@@ -5,11 +5,11 @@ import "time"
 // BlockInfo represents the information a block contains
 type BlockInfo struct {
 	ID           int32     `json:"id" sql:",pk"`
-	Height       int32     `json:"height" sql:",notnull"`
+	Height       int64     `json:"height" sql:",notnull"`
 	Proposer     string    `json:"proposer" sql:",notnull"`
 	BlockHash    string    `json:"block_hash" sql:",notnull,unique"`
 	NumPrecommit string    `json:"num_pre_commit" sql:",notnull"`
-	NumTxs       int32     `json:"num_txs" sql:"default:0"`
-	TotalTxs     int32     `json:"total_txs" sql:"default:0"`
+	NumTxs       int64     `json:"num_txs" sql:"default:0"`
+	TotalTxs     int64     `json:"total_txs" sql:"default:0"`
 	Timestamp    time.Time `json:"timestamp" sql:"default:now()"`
 }
